Feat: new files, folders and features
It was created new alias in the file tsconfig. It was created and implemented new middlewares in the file setupServer. It was created new file named routes for expose the routes. It was created joiValidation. It was created a global variable in express for authentication. It was created a model schema for user, also was created their interfaces. It was created a model schema for video and theirs interfaces too