Feat: new files, folders and features
It was created a new model schema for comments also its interface. It was created validation schemes for signing and signup in the folder schemes in user models. It was a created a new folder named generator and inside of it a new file generators.ts for save functions that will be implemented around the project. The last thing was a created a new environment variable in the file .env and was added to the configEnv too