Feat: new files, folders
It was created a new file named signup.utility.ts in the folder utilities inside of controllers of the model User. Also was create another file named user.service for the configuration of the db. The last thing was a create a new environment variable for JWT in the file .env and this variable was implemented in the file configEnv