Feat: new files for routes and testing endpoints
It was created a new folder and file for routes of the collection user, also this routes was added in the file routes.ts of the folder interface/http . It was created a new folder named endpoints and inside of it was created a new file named auth.http for do testing the endpoints.  It was modified a feature in the controller signup.ts. The lasting was create a new environment variable for test endpoints