Feat: setting jest , new files jest and test
It was created a new file for the jest configuration, also was add a new script in the package.json for jest. It was create a new folder and new file named mocks. The last thing was create a new file named signup.test.ts in the folder controllers of the model user